# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IndoLegalBench API",
    "description": "API IndoLegalBench, platform penulisan dan pengukuran test case hukum untuk Veritask. Dokumentasi ini dihasilkan otomatis dari schema Pydantic dan menjadi kontrak resmi antara server dan client.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/suites": {
      "post": {
        "tags": [
          "suites"
        ],
        "summary": "Buat suite baru",
        "operationId": "create_suite_suites_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuiteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuiteRead"
                }
              }
            }
          },
          "409": {
            "description": "Nama sudah dipakai (`SUITE_NAME_TAKEN`), suite berisi kasus approved (`SUITE_HAS_APPROVED_CASES`), atau suite arsip dihapus.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "suites"
        ],
        "summary": "Daftar suite",
        "operationId": "list_suites_suites_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SuiteStatus",
              "default": "active"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SuiteRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/suites/{suite_id}": {
      "get": {
        "tags": [
          "suites"
        ],
        "summary": "Detail satu suite",
        "operationId": "get_suite_suites__suite_id__get",
        "parameters": [
          {
            "name": "suite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Suite Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuiteRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "suites"
        ],
        "summary": "Ubah suite",
        "operationId": "update_suite_suites__suite_id__patch",
        "parameters": [
          {
            "name": "suite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Suite Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuiteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuiteRead"
                }
              }
            }
          },
          "409": {
            "description": "Nama sudah dipakai (`SUITE_NAME_TAKEN`), suite berisi kasus approved (`SUITE_HAS_APPROVED_CASES`), atau suite arsip dihapus.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "suites"
        ],
        "summary": "Hapus suite, ditolak kalau berisi kasus approved",
        "operationId": "delete_suite_suites__suite_id__delete",
        "parameters": [
          {
            "name": "suite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Suite Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "409": {
            "description": "Nama sudah dipakai (`SUITE_NAME_TAKEN`), suite berisi kasus approved (`SUITE_HAS_APPROVED_CASES`), atau suite arsip dihapus.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/suites/{suite_id}/archive": {
      "post": {
        "tags": [
          "suites"
        ],
        "summary": "Arsipkan suite",
        "operationId": "archive_suite_suites__suite_id__archive_post",
        "parameters": [
          {
            "name": "suite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Suite Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuiteRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/suites/{suite_id}/unarchive": {
      "post": {
        "tags": [
          "suites"
        ],
        "summary": "Aktifkan kembali suite yang diarsipkan",
        "operationId": "unarchive_suite_suites__suite_id__unarchive_post",
        "parameters": [
          {
            "name": "suite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Suite Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuiteRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/suites/{suite_id}/snapshots": {
      "post": {
        "tags": [
          "suites"
        ],
        "summary": "Bekukan suite sebagai snapshot",
        "description": "Freeze the approved cases. No request body and no name.",
        "operationId": "create_snapshot_suites__suite_id__snapshots_post",
        "parameters": [
          {
            "name": "suite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Suite Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotRead"
                }
              }
            }
          },
          "403": {
            "description": "Bukan admin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Suite tidak ditemukan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Suite tidak punya kasus approved (`NOTHING_TO_SNAPSHOT`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "suites"
        ],
        "summary": "Daftar snapshot satu suite",
        "description": "List snapshots, newest first.\n\nAuthor, Reviewer, Admin, and Viewer. SCRUM-138 shows this list to Viewer.\nCreating a snapshot stays Admin only.",
        "operationId": "list_snapshots_suites__suite_id__snapshots_get",
        "parameters": [
          {
            "name": "suite_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Suite Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SnapshotSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/snapshots/{snapshot_id}": {
      "get": {
        "tags": [
          "snapshots"
        ],
        "summary": "Isi satu snapshot",
        "description": "Return the frozen items. The body is the stored copy.\n\nAuthor, Reviewer, Admin, and Viewer. SCRUM-138 shows this page to Viewer.\nCreating a snapshot stays Admin only.",
        "operationId": "get_snapshot_snapshots__snapshot_id__get",
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotRead"
                }
              }
            }
          },
          "404": {
            "description": "Snapshot tidak ditemukan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActorRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ActorRead",
        "description": "Who wrote a version, or who froze a suite snapshot."
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "examples": [
              "No platform account is mapped to this identity."
            ]
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field",
            "description": "Lokasi field yang gagal, misalnya legal_refs[0].pasal"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_SnapshotSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SnapshotSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "title": "Page[SnapshotSummary]"
      },
      "Page_SuiteRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SuiteRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "title": "Page[SuiteRead]"
      },
...
      "SnapshotItemRead": {
        "properties": {
          "case_id": {
            "type": "string",
            "format": "uuid",
            "title": "Case Id"
          },
          "case_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Case Version Id"
          },
          "body": {
            "additionalProperties": true,
            "type": "object",
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "case_version_id",
          "body"
        ],
        "title": "SnapshotItemRead",
        "description": "One frozen case. body is the copy taken at snapshot time."
      },
      "SnapshotRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "suite_id": {
            "type": "string",
            "format": "uuid",
            "title": "Suite Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "author": {
            "$ref": "#/components/schemas/ActorRead"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SnapshotItemRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "id",
          "suite_id",
          "created_at",
          "author",
          "items"
        ],
        "title": "SnapshotRead",
        "description": "GET /snapshots/{id} and the body of POST /suites/{id}/snapshots."
      },
      "SnapshotSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "author": {
            "$ref": "#/components/schemas/ActorRead"
          },
          "case_count": {
            "type": "integer",
            "title": "Case Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "author",
          "case_count"
        ],
        "title": "SnapshotSummary",
        "description": "One row of GET /suites/{id}/snapshots. Items are on the detail route."
      },
...
      "SuiteCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "description": "Nama suite, harus unik"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Tema atau deskripsi singkat suite"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SuiteCreate"
      },
      "SuiteRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "case_count": {
            "type": "integer",
            "title": "Case Count",
            "description": "Jumlah kasus di dalam suite ini"
          },
          "is_empty": {
            "type": "boolean",
            "title": "Is Empty",
            "description": "True kalau suite belum punya kasus sama sekali"
          },
          "exportable": {
            "type": "boolean",
            "title": "Exportable",
            "description": "False kalau suite kosong atau diarsipkan, jadi tidak ikut pengukuran"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "status",
          "case_count",
          "is_empty",
          "exportable",
          "created_at",
          "updated_at"
        ],
        "title": "SuiteRead"
      },
      "SuiteStatus": {
        "type": "string",
        "enum": [
          "active",
          "archived"
        ],
        "title": "SuiteStatus"
      },
      "SuiteUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Null mengosongkan deskripsi. Kalau field ini tidak dikirim, deskripsi tidak berubah."
          }
        },
        "type": "object",
        "title": "SuiteUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}